+++ app/src/main/java/polyadapter/sample/ActivityScope.kt
@@ -1,6 +1,0 @@
-package polyadapter.sample
-
-import javax.inject.Scope
-
-@Scope
-annotation class ActivityScope